Revert "CAT"
This reverts commit d551213b8d38ff6e73e921663b678ea8a695ba91.
