+++ crates/chia-sdk-driver/src/child_from_spend.rs
@@ -1,13 +1,12 @@
-use std::string::ParseError;
-
 use chia_protocol::CoinSpend;
 use clvmr::Allocator;
+use thiserror::Error;
 
 // given a spend, will return info about the coin being spent
-pub trait FromSpend<R, A> {
+pub trait ChildFromSpend<R, A> {
     fn from_spend(
         allocator: &mut Allocator,
         cs: &CoinSpend,
         additional_info: A,
-    ) -> Result<R, ParseError>;
+    ) -> Result<R, FromSpendError>;
 }
