+++ crates/chia-sdk-driver/src/lib.rs
@@ -1,18 +1,16 @@
+mod child_from_spend;
 mod conditions;
-mod from_puzzle;
 mod from_spend;
 mod parse_error;
-mod puzzle;
 mod puzzles;
 mod spend;
 mod spend_context;
 mod spend_error;
 
+pub use child_from_spend::*;
 pub use conditions::*;
-pub use from_puzzle::*;
 pub use from_spend::*;
 pub use parse_error::*;
-pub use puzzle::*;
 pub use puzzles::*;
 pub use spend::*;
 pub use spend_context::*;
