+++ crates/chia-sdk-driver/src/parse_error.rs
@@ -35,7 +35,4 @@ pub enum ParseError {
 
     #[error("mismatched singleton output (maybe no spend revealed the new singleton state)")]
     MismatchedOutput,
-
-    #[error("mismatched layers (e.g., you tried to parse a singleton as a CAT)")]
-    LayerMismatch,
 }
