+++ crates/chia-sdk-driver/src/puzzles/cat.rs
@@ -1,7 +1,5 @@
-mod cat;
 mod cat_spend;
 mod issue_cat;
 
-pub use cat::*;
 pub use cat_spend::*;
 pub use issue_cat::*;
